pyproject.toml
[project]
name = "openrouter-agent-python"
version = "0.1.0"
description = "A powerful CLI-based AI coding agent using OpenRouter (Python Implementation)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "litestar>=2.12.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.7.0",
    "msgspec>=0.18.6",
    "python-socketio>=5.11.0",
    "rich>=13.7.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.1",
    "tenacity>=8.3.0",
    "requests>=2.31.0",
    "jsonpath-ng>=1.6.0",
]

[project.scripts]
openrouter-agent = "python_impl.demo.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openrouter_agent"]
